Updating the Create endpoint to return a 201 (Task 5)
Opted to use CreatedAtRoute since it's low maintenance (will not need manually updated if the endpoint changes) and will generate a URI on the fly rather than having to build one. Due to this, I needed to added a name to the endpoint found within GetById. All data will be prepopulated so that the URI returned will take you to the project you've just created

Updated Swagger with all return types that you can encounter (a mix of tasks 5 and 8)

Added a new ShouldBeCreated method to HttpResponseMessageUtils so that we can use it during testing, as there was no support in place for 201 status codes

Update create tests to ensure that a 201 status code is returned when valid